moved `app` and `db` to __init__ added basic tests fixed PUT status code returned for invalid refresh token (200 -> 400)
